+++ packages/gatsby-plugin-search/.storybook/decorators/react-heading-levels.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { HeadingLevelProvider } from "@jfrk/react-heading-levels";
-
-export function withReactHeadingLevels(storyFn) {
-  return <HeadingLevelProvider>{storyFn()}</HeadingLevelProvider>;
-}