+++ packages/gatsby-plugin-search/.storybook/decorators/react-id.js
@@ -1,6 +1,0 @@
-import React from "react";
-import { IDContextProvider } from "@jfrk/react-id";
-
-export function withReactIDContext(storyFn) {
-  return <IDContextProvider>{storyFn()}</IDContextProvider>;
-}