+++ packages/gatsby-plugin-search/src/utils/ensureArray.js
@@ -1,0 +1,9 @@
+export default function ensureArray(value, isEmpty = (value) => value === "") {
+  if (isEmpty(value)) {
+    return [];
+  }
+  if (Array.isArray(value)) {
+    return value;
+  }
+  return [value];
+}