+++ packages/gatsby-theme-wordpress-basic/src/components/Archive.js
@@ -1,4 +1,5 @@
 import { H, Section } from "@jfrk/react-heading-levels";
+import { withComponentDefaults } from "@whitespace/components";
 import {
   LazyMinisearchSearchBackendProvider,
   SearchContextDebug,
@@ -8,9 +9,9 @@ import {
   SearchPagination,
 } from "@whitespace/gatsby-plugin-search";
 import clsx from "clsx";
+import { flow } from "lodash-es";
 import PropTypes from "prop-types";
 import React from "react";
-import * as yup from "yup";
 
 import {
   getMainArchivePageTitleFromPageContext,
@@ -19,20 +20,25 @@ import {
   getArchiveURLPatternFromPageContext,
 } from "../contentType";
 import { layout } from "../foundation";
-import { usePageContext } from "../hooks";
+import { useArchiveParamTypes, usePageContext } from "../hooks";
 
 import * as defaultStyles from "./Archive.module.css";
 
 Archive.propTypes = {
-  styles: PropTypes.objectOf(PropTypes.string),
   className: PropTypes.string,
+  styles: PropTypes.objectOf(PropTypes.string),
+  transformParams: PropTypes.func,
 };
 
-export default function Archive({
-  styles = defaultStyles,
+export default withComponentDefaults(Archive, "archive");
+
+function Archive({
   className,
+  styles = defaultStyles,
+  transformParams = (params) => params,
   ...restProps
 }) {
+  const paramTypes = useArchiveParamTypes();
   let pageContext = usePageContext();
 
   const forcedParams = {
@@ -70,25 +76,51 @@ export default function Archive({
                     default:
                       return value;
                   }
-                }}
-                schema={yup.object({
-                  // tags: yup.array().default([]),
-                  year: yup.string().ensure(),
-                  month: yup
-                    .string()
-                    .ensure()
-                    .when("year", (year, schema) =>
-                      year ? schema : schema.strip(),
-                    ),
+
+                paramTypes={paramTypes}
+                decodeParams={({ year, month, ...params }) => ({
+                  ...params,
+                  date: month ? `${year}-${month}` : year,
+                })}
+                encodeParams={({ date, ...params }) => ({
+                  ...params,
+                  ...(/^\d{4}$/.test(date) && { year: date }),
+                  ...(/^\d{4}-\d{2}$/.test(date) && {
+                    year: date.substring(0, 4),
+                    month: date.substring(5, 7),
+                  }),
                 })}
               >
                 <LazyMinisearchSearchBackendProvider
                   preload={true}
                   settings={{
-                    attributesForFaceting: ["year", "month"],
+                    attributesForFaceting: ["contentType", "tags", "month"],
                   }}
+                  transformParams={flow([
+                    transformParams,
+                    ({ date, ...params }) => ({
+                      ...params,
+                      ...(/^\d{4}$/.test(date) && {
+                        month: [
+                          "01",
+                          "02",
+                          "03",
+                          "04",
+                          "05",
+                          "06",
+                          "07",
+                          "08",
+                          "09",
+                          "10",
+                          "11",
+                          "12",
+                        ].map((m) => `${date}-${m}`),
+                      }),
+                      ...(/^\d{4}-\d{2}$/.test(date) && { month: [date] }),
+                    }),
+                  ])}
                 >
-                  <SearchForm showHitsTotal={false} />
+                  <SearchForm />
                   {process.env.NODE_ENV !== "production" && (
                     <details>
                       <summary>Debug</summary>