+++ packages/gatsby-theme-wordpress-basic/src/components/InternalLinkElement.js
@@ -4,16 +4,6 @@ import React from "react";
 
 InternalLinkElement.propTypes = { href: PropTypes.any };
 
-function InternalLinkElement({ href, ...restProps }) {
+export default function InternalLinkElement({ href, ...restProps }) {
   return <GatsbyLink to={href} {...restProps} />;
 }
-
-export default {
-  default: {},
-  breadcrumbs: {},
-  link: {
-    components: {
-      InternalLinkElement,
-    },
-  },
-};