+++ packages/gatsby-theme-wordpress-basic/src/theme/index.js
@@ -1,0 +1,9 @@
+import InternalLinkElement from "../components/InternalLinkElement";
+
+export default {
+  link: {
+    components: {
+      InternalLinkElement,
+    },
+  },
+};