+++ packages/gatsby-theme-wordpress-basic/src/utils/formatMonth.js
@@ -1,0 +1,10 @@
+import capitalizeFirstLetter from "./capitalizeFirstLetter";
+
+export default function formatMonth(date, locale) {
+  return capitalizeFirstLetter(
+    date.toLocaleDateString(locale, {
+      month: "long",
+    }),
+    locale,
+  );
+}