+++ packages/gatsby-theme-wordpress-basic/src/utils/index.js
@@ -1,0 +1,4 @@
+export { default as capitalizeFirstLetter } from "./capitalizeFirstLetter";
+export { default as formatMonth } from "./formatMonth";
+export { default as fromFacetsToOptions } from "./fromFacetsToOptions";
+export { default as getOptionsFromTaxonomy } from "./getOptionsFromTaxonomy";