+++ packages/gatsby-plugin-cookie-consent/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.1.3](https://github.com/whitespace-se/gatsby-packages/compare/@whitespace/gatsby-plugin-cookie-consent@0.1.2...@whitespace/gatsby-plugin-cookie-consent@0.1.3) (2021-10-22)
+
+
+### Bug Fixes
+
+* avoid error on empty whitelist ([99907d0](https://github.com/whitespace-se/gatsby-packages/commit/99907d06a8f2c2fb221f2b28deceb1ff379680e3))
+
+
+
 ### [0.1.2](https://github.com/whitespace-se/gatsby-packages/compare/@whitespace/gatsby-plugin-cookie-consent@0.1.1...@whitespace/gatsby-plugin-cookie-consent@0.1.2) (2021-10-08)
 
 **Note:** Version bump only for package @whitespace/gatsby-plugin-cookie-consent